Moved the function of sending email into a separate file
